refactor: move core ast code to separate module
