+++ apps/language_server/lib/language_server/providers/execute_command/manipulate_pipes.ex
@@ -9,6 +9,8 @@ defmodule ElixirLS.LanguageServer.Providers.ExecuteCommand.ManipulatePipes do
 
   alias ElixirLS.LanguageServer.{JsonRpc, Server}
 
+  alias __MODULE__.AST
+
   @behaviour ElixirLS.LanguageServer.Providers.ExecuteCommand
 
   @impl ElixirLS.LanguageServer.Providers.ExecuteCommand
@@ -81,7 +83,7 @@ defmodule ElixirLS.LanguageServer.Providers.ExecuteCommand.ManipulatePipes do
         piped_text =
           function_call
           |> Code.string_to_quoted!()
-          |> to_pipe()
+          |> AST.to_pipe()
           |> Macro.to_string()
 
         {:ok, %{edited_text: piped_text, edit_range: range}}
@@ -145,36 +147,5 @@ defmodule ElixirLS.LanguageServer.Providers.ExecuteCommand.ManipulatePipes do
     end)
     |> Map.get(:text)
     |> IO.iodata_to_binary()
-  end
-
-  @doc false
-  def to_pipe(code) do
-    {piped_ast, _} = Macro.prewalk(code, %{has_piped: false}, &to_pipe/2)
-    piped_ast
-  end
-
-  defp to_pipe({:|>, line, [left, right]}, %{has_piped: false} = acc) do
-    {{:|>, line, [left |> to_pipe(acc) |> elem(0), right]}, Map.put(acc, :has_piped, true)}
-  end
-
-  defp to_pipe(
-         {{:., line, [{_, _, nil}]} = anonymous_function_node, _meta, [h | t]},
-         %{has_piped: false} = acc
-       ) do
-    {{:|>, line, [h, {anonymous_function_node, line, t}]}, Map.put(acc, :has_piped, true)}
-  end
-
-  defp to_pipe({{:., line, _args} = function, _meta, args}, %{has_piped: false} = acc)
-       when args != [] do
-    {{:|>, line, [hd(args), {function, line, tl(args)}]}, Map.put(acc, :has_piped, true)}
-  end
-
-  defp to_pipe({function, line, [h | t]}, %{has_piped: false} = acc)
-       when is_atom(function) and t != [] do
-    {{:|>, line, [h, {function, line, t}]}, Map.put(acc, :has_piped, true)}
-  end
-
-  defp to_pipe(node, acc) do
-    {node, acc}
   end
 end
